pyproject.toml
[tool.poetry]
name = "jw_test"
version = "0.1.0"
description = ""
authors = ["alfa24"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
django-model-utils = "^4.1.1"
celery = "^5.1.2"
psycopg2-binary = "^2.9.1"
lorem = "^0.1.1"
django-admin-sortable2 = "^1.0"
redis = "^3.5.3"
mixer = "^7.1.2"
Faker = "^8.12.1"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.6.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
